refactor(Hotkey): move hotkey to common
